+++ utils/remote_retrieve.sh
@@ -11,7 +11,7 @@ mkdir -p ../backups
 scp -C -P "${remote_port}" "${remote_username}@${remote_host}:$remote_m2_db_file" ../backups/latest-m2.sql.gz
 gunzip -f ../backups/latest-m2.sql.gz
 
-if existRemoteFile "$remote_wp_db_file"; then
+if [ $_arg_wordpress == 'on' ]; then
     scp -C -P "${remote_port}" "${remote_username}@${remote_host}:$remote_backup_dir/latest-wp.sql.gz" ../backups/latest-wp.sql.gz
     gunzip -f ../backups/latest-wp.sql.gz
 fi